+++ docs/resources/changelog.md
@@ -15,8 +15,9 @@ _During the beta period, these restrictions may be relaxed in the event of a mis
 - Adjusted `--sl-overlay-background-color` in dark theme to be black instead of gray
 - Fixed a bug in `sl-color-picker` where the opacity slider wasn't showing the current color
 - Fixed a bug where Edge in Windows would show the native password toggle next to the custom password toggle [#508](https://github.com/shoelace-style/shoelace/issues/508)
-- Improved color picker size
-- Improved input icon contrast
+- Improved size of `sl-color-picker`
+- Improved icon contrast in `sl-input`
+- Improved contrast of `sl-switch`
 - Removed elevation from `sl-color-picker` when rendered inline
 - Updated to Lit 2.0.0-rc.3
 - Updated to lit-html 2.0.0-rc.4